API

Modules

Types and constants

LegendDataManagement.AbstractSetupData

LegendDataManagement.AnyProps

LegendDataManagement.LegendData

LegendDataManagement.LegendDataConfig

LegendDataManagement.LegendTierData

LegendDataManagement.PropsDB

LegendDataManagement.SetupConfig

LegendDataManagement.ValiditySelection

Functions and macros

LegendDataManagement.channel_info

LegendDataManagement.get_setup_config

LegendDataManagement.setup_data_path

LegendDataManagement.setup_data_path

Documentation

LegendDataManagement.AbstractSetupData — Type

abstract type AbstractSetupData

Subtypes wrap SetupConfig for specific experiments.

LegendDataManagement.LegendData — Type

struct LegendData <: AbstractSetupData

Provides access to LEGEND data and metadata.

Constructors:

LegendData(setup_config::SetupConfig).

LegendData(setup::Symbol) - requires the $LEGEND_DATA_CONFIG environment variable to be set.

Examples:

config_filename = "/path/to/config.json"
config = LegendDataConfig(config_filename)
l200 = LegendData(config.setups.l200)

filekey = FileKey("l200-p02-r006-cal-20221226T200846Z")

or simply (if $LEGEND_DATA_CONFIG is set):

l200 = LegendData(:l200)

LegendData has the (virtual) properties metadata and tier.

The full path to "tier" data files can be retrieved using

(data::LegendData)[tier::Symbol, filekey::FileKey]
(data::LegendData).tier[tier::Symbol, filekey::AbstractString]

Example:

l200.tier.dsp[:raw, "l200-p02-r006-cal-20221226T200846Z"]
l200.tier.dsp("l200-p02-r006-cal-20221226T200846Z"]

LegendDataManagement.LegendDataConfig — Type

struct LegendDataConfig

Data configuration multiple experimental setups.

Contains a single field setups::PropertyDict{Symbol,SetupConfig}.

Can be read from a config file via LegendDataConfig(config_filename), or simply LegendDataConfig() if the environment variable $LEGEND_DATA_CONFIG is set.

Example:

config = LegendDataConfig("/path/to/config.json")
setup = config.setups.l200
setup_data_path(setup, ["tier", "raw", "cal", "p02", "r006", "l200-p02-r006-cal-20221226T200846Z-tier_raw.lh5"])

See also SetupConfig.

LegendDataManagement.LegendTierData — Type

struct LegendDataManagement.LegendTierData

Constructors:

(data::LegendData).tier

LegendDataManagement.LegendTierData(data::LegendData)

tier_data::LegendTierData supports

The full path of data files can be retrieved using

data[tier::Symbol, filekey::FileKey]
data[tier::Symbol, filekey::AbstractString]

LegendDataManagement.PropsDB — Type

struct LegendDataManagement.PropsDB

Use code should not instantiate PropsDB directly, use LegendDataManagement.AnyProps(base_path::AbstractString) instead.

LegendDataManagement.SetupConfig — Type

struct SetupConfig

Data configuration for an experimental setup.

Supports

setup_data_path(setup, path_components)

Examples:

setup_data_path(setup, ["tier", "raw", "cal", "p02", "r006", "l200-p02-r006-cal-20221226T200846Z-tier_raw.lh5"])

LegendDataManagement.ValiditySelection — Type

struct LegendDataManagement.ValiditySelection

Representy validiy selection for a LegendDataManagement.PropsDB[@ref].

LegendDataManagement.AnyProps — Type

LegendDataManagement.AnyProps = Union{LegendDataManagement.PropsDB,PropDicts.PropDict}

Properties stored either in a directory managed via [LegendDataManagement.PropsDB][@ref] or loaded from one or several files into a PropDicts.PropDict.

Constructors:

LegendDataManagement.AnyProps(base_path::AbstractString)

LegendDataManagement.channel_info — Method

channel_info(data::LegendData, filekey::FileKey)

Get channel information for a given filekey.

LegendDataManagement.get_setup_config — Function

LegendDataManagement.get_setup_config(data::AbstractSetupData)::SetupConfig

Must be specialized for each subtype of AbstractSetupData.

LegendDataManagement.setup_data_path — Function

setup_data_path(setup::SetupConfig, path_components::AbstractVector{<:AbstractString})
setup_data_path(setup::SetupConfig, path::AbstractString})

Get the full absolute path for the given path_components as configured for setup.

LegendDataManagement.setup_data_path — Method

setup_data_path(setup::AbstractSetupData, path_components::AbstractVector{<:AbstractString})
setup_data_path(setup::AbstractSetupData, path::AbstractString})

Get the full absolute path for the given path_components as configured for setup.