API

Modules

Types and constants

LegendDataManagement.AbstractSetupData

LegendDataManagement.AnyProps

LegendDataManagement.DataCategory

LegendDataManagement.DataCategoryLike

LegendDataManagement.DataPeriod

LegendDataManagement.DataPeriodLike

LegendDataManagement.DataRun

LegendDataManagement.DataRunLike

LegendDataManagement.DataSelector

LegendDataManagement.DataTier

LegendDataManagement.DataTierLike

LegendDataManagement.ExpSetup

LegendDataManagement.ExpSetupLike

LegendDataManagement.FileKey

LegendDataManagement.FileKeyLike

LegendDataManagement.LegendData

LegendDataManagement.LegendDataConfig

LegendDataManagement.LegendTierData

LegendDataManagement.PropsDB

LegendDataManagement.SetupConfig

LegendDataManagement.Timestamp

LegendDataManagement.TimestampLike

LegendDataManagement.ValiditySelection

Functions and macros

LegendDataManagement.channel_info

LegendDataManagement.data_path

LegendDataManagement.data_path

LegendDataManagement.data_path

LegendDataManagement.data_path

LegendDataManagement.get_setup_config

LegendDataManagement.legend_addprocs

LegendDataManagement.read_filekeys

LegendDataManagement.search_disk

LegendDataManagement.write_filekeys

Documentation

LegendDataManagement.AbstractSetupData — Type

abstract type AbstractSetupData

Subtypes wrap SetupConfig for specific experiments.

LegendDataManagement.DataCategory — Type

struct DataCategory <: DataSelector

Represents a LEGEND data category (related to a DAQ/measuring mode) like "cal" or "phy".

Example:

category = DataCategory(:cal)
category.label == :cal
string(category) == "cal"
DataCategory("cal") == category

LegendDataManagement.DataPeriod — Type

struct DataPeriod <: DataSelector

Represents a LEGEND data-taking period.

Example:

period = DataPeriod(2)
period.no == 2
string(period) == "p02"
DataPeriod("p02") == period

LegendDataManagement.DataRun — Type

struct DataRun <: DataSelector

Represents a LEGEND data-taking run.

Example:

```julia r = DataRun(6) r.no == 6 string(r) == "r006" DataRun("r006") == r

LegendDataManagement.DataSelector — Type

abstract type DataSelector

Abstract type for data selectors like ExpSetup, DataTier, DataPeriod, DataRun, DataCategory, Timestamp and FileKey.

LegendDataManagement.DataTier — Type

struct DataTier <: DataSelector

Represents a LEGEND data tier like "raw, "dsp", etc.

Example:

tier = DataTier(:raw)
tier.label == :raw
string(tier) == "raw"
DataTier("raw") == tier

LegendDataManagement.ExpSetup — Type

struct ExpSetup <: DataSelector

Represents a LEGEND experimental setup like "l200".

Example:

setup = ExpSetup(:l200)
setup.label == :l200
string(setup) == "l200"
ExpSetup("l200") == setup

LegendDataManagement.FileKey — Type

struct FileKey <: DataSelector

Represents a LEGEND file key.

Example:

filekey = FileKey("l200-p02-r006-cal-20221226T200846Z")

See also read_filekeys and write_filekeys.

LegendDataManagement.LegendData — Type

struct LegendData <: AbstractSetupData

Provides access to LEGEND data and metadata.

Constructors:

LegendData(setup_config::SetupConfig).

LegendData(setup::Symbol) - requires the $LEGEND_DATA_CONFIG environment variable to be set.

Examples:

config_filename = "/path/to/config.json"
config = LegendDataConfig(config_filename)
l200 = LegendData(config.setups.l200)

filekey = FileKey("l200-p02-r006-cal-20221226T200846Z")

or simply (if $LEGEND_DATA_CONFIG is set):

l200 = LegendData(:l200)

LegendData has the (virtual) properties metadata and tier.

The full path to "tier" data files can be retrieved using

(data::LegendData)[tier::Symbol, filekey::FileKey]
(data::LegendData).tier[tier::Symbol, filekey::AbstractString]

Example:

l200.tier[:raw]
l200.tier[:raw, FileKey("l200-p02-r006-cal-20221226T200846Z")]

LegendData comes with an extension for SolidStateDetectors:

l200 = LegendData(:l200)
SolidStateDetector(l200, :V99000A)

LegendDataManagement.LegendDataConfig — Type

struct LegendDataConfig

Data configuration multiple experimental setups.

Contains a single field setups::PropertyDict{Symbol,SetupConfig}.

Can be read from a config file via LegendDataConfig(config_filename[s]), or simply LegendDataConfig() if the environment variable $LEGEND_DATA_CONFIG is set. $LEGEND_DATA_CONFIG may be a list of colon-separated config filenames, which are applied/merged in reverse order (analog to the order of prioritiy in $PATH and similar).

Example:

config = LegendDataConfig("/path/to/config.json")
setup = config.setups.l200
data_path(setup, "tier", "raw", "cal", "p02", "r006", "l200-p02-r006-cal-20221226T200846Z-tier_raw.lh5")

See also SetupConfig.

LegendDataManagement.LegendTierData — Type

struct LegendDataManagement.LegendTierData

Constructors:

(data::LegendData).tier

LegendDataManagement.LegendTierData(data::LegendData)

The path to data directories and files can be accessed via getindex on tier_data::LegendTierData:

tier_data[]
tier_data[tier::DataTierLike]
tier_data[tier::DataTierLike, category::DataCategoryLike]
tier_data[tier::DataTierLike, category::DataCategoryLike, period::DataPeriodLike]
tier_data[tier::DataTierLike, category::DataCategoryLike, period::DataPeriodLike, run::DataRunLike]

tier_data[tier::DataTierLike, filekey::FileKeyLike]

Examples:

```julia l200 = LegendData(:l200)

filekey = FileKey("l200-p02-r006-cal-20221226T200846Z") isfile(l200.tier[:raw, filekey])

isdir(l200.tier[:raw, :cal]) isdir(l200.tier[:raw, :cal, "p02"]) isdir(l200.tier[:raw, :cal, "p02", "r006"]) isdir(l200.tier[DataTier(:raw), DataCategory(:cal), DataPeriod(2), DataRun(6)])

LegendDataManagement.PropsDB — Type

struct LegendDataManagement.PropsDB

A PropsDB instance, e.g. myprops, presents an on-disk directory containing JSON files or sub-directories (that contains JSON files in leaf directories) as a dictionary of properties.

PropsDB supports Base.keys and Base.getindex as well as Base.propertynames and Base.getproperty to access it's contents. getindex and getproperty will return either another PropsDB or a PropDicts.PropDict, depending on whether the accessed property is stored as a sub-directory or a JSON file. We recommend to use getproperty where the properties/keys of the PropDict are more or less standardized and where they may be arbitrary (see examples below).

The contents of PropsDB may be time- and category-dependent, determined by the presence of a "validity.json" file. In this case, use myprops(sel::LegendDataManagement.ValiditySelection) or myprops(filekey::FileKey) to select the desired time and category. The selection can be made at some point during traversal of properties or at the leaf PropsDB (see the examples below).

Examples:

l200 = LegendData(:l200)

propertynames(l200.metadata.hardware)
l200.metadata.hardware.detectors.germanium

keys(l200.metadata.hardware.detectors.germanium.diodes)
l200.metadata.hardware.detectors.germanium.diodes[:V99000A]

diodes = l200.metadata.hardware.detectors.germanium.diodes
diodes[keys(diodes)]

sel = ValiditySelection("20221226T194007Z", :cal)
filekey = FileKey("l200-p02-r006-cal-20221226T194007Z")
data.metadata.hardware(sel).configuration.channelmaps
data.metadata.hardware.configuration.channelmaps(filekey)

Use code should not instantiate PropsDB directly, use LegendDataManagement.AnyProps(path::AbstractString) instead, which may return a PropsDB or a PropDicts.PropDict depending on what on-disk content path points to.

LegendDataManagement.SetupConfig — Type

struct SetupConfig

Data configuration for an experimental setup.

Supports

data_path(setup, path_components)

Examples:

data_path(setup, "tier", "raw", "cal", "p02", "r006", "l200-p02-r006-cal-20221226T200846Z-tier_raw.lh5")

LegendDataManagement.Timestamp — Type

struct Timestamp <: DataSelector

Represents a LEGEND timestamp.

Example:

julia timestamp = Timestamp("20221226T200846Z") timestamp.unixtime == 1672085326 string(timestamp) == "20221226T200846Z"`

LegendDataManagement.ValiditySelection — Type

struct LegendDataManagement.ValiditySelection

Representy validiy selection for a LegendDataManagement.PropsDB[@ref].

LegendDataManagement.AnyProps — Type

LegendDataManagement.AnyProps = Union{LegendDataManagement.PropsDB,PropDicts.PropDict}

Properties stored either in a directory managed via [LegendDataManagement.PropsDB][@ref] or loaded from one or several files into a PropDicts.PropDict.

Constructors:

LegendDataManagement.AnyProps(base_path::AbstractString)

LegendDataManagement.DataCategoryLike — Type

DataCategoryLike = Union{DataCategory, Symbol, AbstractString}

Anything that can represent a data category, like DataCategory(:cal), :cal or "cal".

LegendDataManagement.DataPeriodLike — Type

DataPeriodLike = Union{DataPeriod, Integer, AbstractString}

Anything that can represent a data period, like DataPeriod(2) or "p02".

LegendDataManagement.DataRunLike — Type

DataRunLike = Union{DataRun, Integer, AbstractString}

Anything that can represent a data run, like DataRun(6) or "r006".

LegendDataManagement.DataTierLike — Type

DataTierLike = Union{DataTier, Symbol, AbstractString}

Anything that can represent a data tier, like DataTier(:raw), :raw or "raw".

LegendDataManagement.ExpSetupLike — Type

ExpSetupLike = Union{ExpSetup, Symbol, AbstractString}

Anything that can represent a setup label, like ExpSetup(:l200), :l200 or "l200".

LegendDataManagement.FileKeyLike — Type

FileKeyLike = Union{FileKey, AbstractString}

Anything that can represent a file key, like FileKey("l200-p02-r006-cal-20221226T200846Z") or "l200-p02-r006-cal-20221226T200846Z".

LegendDataManagement.TimestampLike — Type

TimestampLike = Union{Timestamp, AbstractString, Integer}

Anything that can represent a timestamp, like Timestamp("20221226T200846Z") or "20221226T200846Z".

LegendDataManagement.channel_info — Method

channel_info(data::LegendData, filekey::FileKey)

Get channel information for a given filekey.

LegendDataManagement.data_path — Function

data_path(setup::SetupConfig, path_components::AbstractString...)

Get the full absolute path for the given path_components as configured for setup.

Converts between "/" and "\" inside of path_components if necessary (for Windows compatibility).

LegendDataManagement.data_path — Method

data_path(setup::AbstractSetupData, path_components::AbstractString...)

Get the full absolute path for the given path_components as configured for setup.

LegendDataManagement.data_path — Method

data_path(tier_data::LegendTierData, path_components::AbstractString...)

Get the full absolute path for the given path_components relative to tier_data.

LegendDataManagement.data_path — Method

data_path(pd::LegendDataManagement.PropsDB)

Return the path to the data directory that contains pd.

LegendDataManagement.get_setup_config — Function

LegendDataManagement.get_setup_config(data::AbstractSetupData)::SetupConfig

Must be specialized for each subtype of AbstractSetupData.

LegendDataManagement.legend_addprocs — Function

legend_addprocs(nprocs::Integer)

Add Julia workers for LEGEND data processing.

Calls Distributed.addprocs with some specific options.

Ensures that all workers processes use the same Julia project environment as the current process. Requires that file systems paths are consistenst across compute nodes.

LegendDataManagement.read_filekeys — Method

read_filekeys(filename::AbstractString)::AbstractVector{FileKey}

Reads a list of FileKey from a text file, one file key per line.

Ignores empty lines. # may be used to start a comment in the file.

LegendDataManagement.search_disk — Function

search_disk(::Type{<:DataSelector}, path::AbstractString)

Search on-disk data for data categories, periods, runs, and filekeys.

Examples:

l200 = LegendData(:l200)

search_disk(DataCategory, l200.tier[:raw])
search_disk(DataPeriod, l200.tier[:raw, :cal])
search_disk(DataRun, l200.tier[:raw, :cal, "p02"])
search_disk(FileKey, l200.tier[DataTier(:raw), :cal, DataPeriod(2), "r006"])

LegendDataManagement.write_filekeys — Method

write_filekeys(filename::AbstractString, filekeys::AbstractVector{<:FileKey})

Writes a list of FileKey to a text file, one file key per line.